+++ vscode/microsoft-kiota/src/kiotaInterop/index.ts
@@ -24,6 +24,7 @@ export async function connectToKiota<T>(context: vscode.ExtensionContext, callba
     try {
         return await callback(connection);
     } catch (error) {
+        console.warn(error);
         const errorMessage = (error as { data?: { message: string } })?.data?.message
             || 'An unknown error occurred';
         vscode.window.showErrorMessage(errorMessage);
@@ -248,6 +249,13 @@ export interface GenerationConfiguration {
     usesBackingStore: boolean;
     pluginTypes: KiotaPluginType[];
     operation: ConsumerOperation;
+    pluginAuthRefid?: string;
+    pluginAuthType?: PluginAuthType | null;
+}
+
+export enum PluginAuthType {
+    oAuthPluginVault = "OAuthPluginVault",
+    apiKeyPluginVault = "ApiKeyPluginVault"
 }
 
 interface WorkspaceObjectProperties {
@@ -269,6 +277,8 @@ export interface ClientObjectProperties extends WorkspaceObjectProperties {
 
 export interface PluginObjectProperties extends WorkspaceObjectProperties {
     types: string[];
+    authType?: PluginAuthType,
+    authReferenceId?: string;
 }
 
 export type ClientOrPluginProperties = ClientObjectProperties | PluginObjectProperties;