+++ vscode/microsoft-kiota/src/providers/openApiTreeProvider.ts
@@ -353,7 +353,6 @@ export class OpenApiTreeProvider implements vscode.TreeDataProvider<OpenApiTreeN
                         clientNameOrPluginName
                     );
                 }
-                await updateTreeViewIcons(treeViewId, true, false);
             }
         }
     }
@@ -419,7 +418,7 @@ function createKiotaOpenApiNode(
         clientNameOrPluginName
     };
 }
-type IconSet = string | vscode.Uri | { light: string | vscode.Uri; dark: string | vscode.Uri } | vscode.ThemeIcon;
+type IconSet = string | vscode.Uri | vscode.ThemeIcon;
 export class OpenApiTreeNode extends vscode.TreeItem {
     private static readonly selectedSet: IconSet = new vscode.ThemeIcon('check');